requirements.txt
PyOpenGL 
PyOpenGL_accelerate
pygame